feat(core): Support for 64 layers
Ability to use for more than 32 layers by switching layer variable type from uint32_t to uint64_t.

- Add macro: WRITE_BIT64
- Update layer code to use BIT64 and WRITE_BIT64
- Tested and working on Corne Mini

+++ app/include/zmk/keymap.h
@@ -30,7 +30,7 @@ typedef uint8_t zmk_keymap_layer_id_t;
  */
 typedef uint8_t zmk_keymap_layer_index_t;
 
-typedef uint32_t zmk_keymap_layers_state_t;
+typedef uint64_t zmk_keymap_layers_state_t;
 
 zmk_keymap_layer_id_t zmk_keymap_layer_index_to_id(zmk_keymap_layer_index_t layer_index);
